Changed the disabled indicator to be all LEDs blinking
There is still an issue where the controller will go into a kind of
sleep mode if nothing is pressed for an extended period of time.  This
does not seem to adversely affect the responsiveness of commands but
does cause the status LEDs to change to the rotating setting when it
occurs.  Will need to include this in the user manual.